requirements.txt
gunicorn
flask
bcrypt
smtplib
mysql-connector
Wand
Requests